refactor(ai): move stream processor into streaming package
